src/wrapper_scripts/MELD_env_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
anndata=0.7.5=pypi_0
argon2-cffi=20.1.0=py36h1d69622_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
batchglm=0.7.4=pypi_0
bleach=3.2.1=pyh9f0ad1d_0
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.2=pypi_0
certifi=2020.12.5=py36h5fab9bb_0
cffi=1.14.4=py36hc120d54_1
cmocean=2.0=py_3
colorspacious=1.1.2=pyh24bf2e0_0
cycler=0.10.0=py_2
dask=2020.12.0=pypi_0
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
deprecated=1.2.10=pyh9f0ad1d_0
diffxpy=0.7.4=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.2.9=he1b5a44_2
fontconfig=2.13.1=h736d332_1003
freetype=2.10.4=h7ca028e_0
future=0.18.2=py36h5fab9bb_2
gettext=0.19.8.1=h0b5b191_1005
glib=2.66.4=hcd2ae1e_1
graphtools=1.5.2=py_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.18.2=h3560a44_1
h5py=3.1.0=pypi_0
icu=68.1=h58526e2_0
importlib-metadata=3.3.0=py36h5fab9bb_2
importlib_metadata=3.3.0=hd8ed1ab_2
ipykernel=5.4.2=py36he448a4c_0
ipython=7.12.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.2=pyhd3deb0d_1
jedi=0.18.0=py36h5fab9bb_1
jinja2=2.11.2=pyh9f0ad1d_0
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py_2
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.0=py36h5fab9bb_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py36h51d7077_0
krb5=1.17.2=h926e7f8_0
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hea4e1c9_1
libblas=3.9.0=6_openblas
libcblas=3.9.0=6_openblas
libclang=11.0.0=default_ha5c780c_2
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libglib=2.66.4=h164308a_1
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.9.0=6_openblas
libllvm11=11.0.0=he513fc3_0
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libpq=12.3=h255efa7_3
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_3
llvmlite=0.35.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_0
magic-impute=2.0.4=pypi_0
markupsafe=1.1.1=py36he6145b8_2
matplotlib=3.3.3=py36h5fab9bb_0
matplotlib-base=3.3.3=py36he12231b_0
meld=1.0.0=pypi_0
mistune=0.8.4=py36h1d69622_1002
mysql-common=8.0.22=ha770c72_1
mysql-libs=8.0.22=h1fd7589_1
natsort=7.1.0=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py36h5fab9bb_3
nbformat=5.0.8=py_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.4.3=pyhd8ed1ab_0
notebook=6.1.6=py36h5fab9bb_0
nspr=4.29=h9c3ff4c_1
nss=3.60=hb5efdd6_0
numba=0.52.0=pypi_0
numpy=1.19.4=py36h2aa4a07_2
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h7f98852_0
packaging=20.8=pyhd3deb0d_0
pandas=1.1.5=py36h284efc9_0
pandoc=2.11.3.2=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.1=pyhd8ed1ab_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyh9f0ad1d_2
phate=1.0.4=py_0
pickle5=0.0.11=py36h8c4c3a4_0
pickleshare=0.7.5=py_1003
pillow=8.1.0=py36h4f9996e_0
pip=20.3.3=pyhd8ed1ab_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.8=pyha770c72_0
prompt_toolkit=3.0.8=hd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.3=pyhd8ed1ab_0
pygsp=0.5.1=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py36h5fab9bb_6
pyqt-impl=5.12.3=py36h7ec31b9_6
pyqt5-sip=4.19.18=py36hc4f0c31_6
pyqtchart=5.12=py36h7ec31b9_6
pyqtwebengine=5.12.1=py36h7ec31b9_6
pyrsistent=0.17.3=py36h1d69622_1
python=3.6.12=hffdb5ce_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytz=2020.5=pyhd8ed1ab_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=py36h2b1bd32_1
qt=5.12.9=h9d6b050_2
qtconsole=5.0.1=pyhd8ed1ab_0
qtpy=1.9.0=py_0
readline=8.0=he28a2e2_2
s_gd2=1.7=py36hdb11119_2
scikit-learn=0.24.0=py36he4fde30_0
scipy=1.5.3=py36h9e8f40b_0
scprep=1.0.11=py_0
seaborn=0.11.1=ha770c72_0
seaborn-base=0.11.1=pyhd8ed1ab_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py36h9880bd3_2
six=1.15.0=pyh9f0ad1d_0
sklearn=0.0=pypi_0
sparse=0.11.2=pypi_0
sqlite=3.34.0=h74cdb3f_0
statsmodels=0.12.1=py36ha112f06_1
tasklogger=1.0.0=py_0
terminado=0.9.1=py36h5fab9bb_1
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h21135ba_1
toolz=0.11.1=pypi_0
tornado=6.1=py36h1d69622_0
traitlets=4.3.3=py36h9f0ad1d_1
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py36h5fab9bb_4
wrapt=1.12.1=py36h1d69622_2
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h516909a_1
zeromq=4.3.3=h58526e2_3
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.8=ha95c52a_1